requirements.txt
roboflow
torch
opencv-python
numpy
Pillow
easyocr
streamlit
roboflow
seaborn